VLEK: Add ASVK support
With the release of the VLEK, there is a separate api endpoint for
requesting VLEK signing keys which are known as an AMD Signing VLEK Key
(ASVK). As the GHCB does not offer a unique UUID for ASVK, treat ASVK
as ASK to include it in the kernel format file. To export the
certificate chain back from the file, tell ASVK from ASK apart by
checking the existance of VLEK. Assumes the certificate as ASK if VLEK
is missing from the kernel format file.

Signed-off-by: Ian Chin Wang <ian.chin.wang@oracle.com>