Add support to fetch VLEK hashsticks
Add the hashsticks under command fetch. Users may choose to fetch the
hashsticks for the current machine, for a hwID specified by the users,
or for a list of the machines with the same processor generation and
the reported TCB version as the current machine by providing a list.
This command writes the hashsticks to a file called "hashsticks". If
only one hwID is queried for the hashstick, the command also convert
the file to a binary accepted by VLEK load.

Signed-off-by: Ian Chin Wang <ian.chin.wang@oracle.com>